[refactor] Refactor to match the new auto-import
